Rename qwebview -> proxy_qwebview
