+++ jigna/core/proxy_qwebview.py
